+++ packages/analytics-core/src/initialization/browser-adapter.ts
@@ -1,3 +1,4 @@
+import { getBotCookieClientSide, isBot } from '../bot-detection/bot-detection';
 import {
   COOKIE_NAME_PREFIX,
   fetchClientIdFromEdgeProxy,
@@ -30,6 +31,10 @@ export interface AnalyticsBrowserAdapter extends AnalyticsAdapter {
 export function analyticsBrowserAdapter(): AnalyticsBrowserAdapter {
   return {
     type: 'browser',
+    isBot: () => {
+      const botCookie = getBotCookieClientSide();
+      return !!botCookie || isBot(navigator.userAgent);
+    },
     getClientId: () => {
       return getCookieValueClientSide(getAnalyticsPlugin().options.cookies.name);
     },