+++ packages/nextjs/src/proxy/personalize-proxy.ts
@@ -6,9 +6,9 @@ import {
   CdpHelper,
   DEFAULT_VARIANT,
 } from '@sitecore-content-sdk/content/personalize';
+import { BOT_DETECTION_COOKIE } from '@sitecore-content-sdk/analytics-core/internal';
 import { initContentSdk } from '@sitecore-content-sdk/core';
 import { personalize } from '@sitecore-content-sdk/personalize';
-import { BOT_DETECTION_COOKIE } from '@sitecore-content-sdk/events/internal';
 import { analyticsPlugin } from '@sitecore-content-sdk/analytics-core';
 import { personalizeServerPlugin } from '@sitecore-content-sdk/personalize';
 import { analyticsProxyAdapter } from '../initialization/proxy/analytics-adapter';